sessions path building

+++ frontend_fleet/src/app/Common-components/map/map.component.ts
@@ -14,6 +14,7 @@ import { CommonModule } from '@angular/common';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { Realtime } from '../../Models/Realtime';
 import { RealtimeData } from '../../Models/RealtimeData';
+import { Point } from '../../Models/Point';
 
 @Component({
   selector: 'app-map',
@@ -29,10 +30,6 @@ export class MapComponent implements AfterViewInit, OnDestroy {
   map!: L.Map;
   initialized: boolean = false;
 
-  //parametri default per Perugia
-  private lat: number = 43.1121;
-  private long: number = 12.3888;
-
   constructor(private mapService: MapService, private cd: ChangeDetectorRef) {}
 
   ngOnDestroy(): void {
@@ -40,18 +37,19 @@ export class MapComponent implements AfterViewInit, OnDestroy {
     this.destroy$.complete();
   }
 
+  private routingControl: L.Routing.Control | null = null;
+
   ngAfterViewInit(): void {
     this.mapService.loadMap$.pipe(takeUntil(this.destroy$), skip(1)).subscribe({
       next: (realtimeData: RealtimeData | null) => {
         if (realtimeData && realtimeData.realtime) {
           console.log(realtimeData.anomaly);
-          this.lat = realtimeData.realtime.latitude;
-          this.long = realtimeData.realtime.longitude;
-          console.log();
-          this.initMap();
+          const pgLat = realtimeData.realtime.latitude;
+          const pgLong = realtimeData.realtime.longitude;
+          this.initMap(new Point(pgLat, pgLong));
           const marker = this.mapService.createMarker(
-            this.lat,
-            this.long,
+            pgLat,
+            pgLong,
             realtimeData.vehicle.plate,
             realtimeData.anomaly
           );
@@ -59,12 +57,84 @@ export class MapComponent implements AfterViewInit, OnDestroy {
         }
       },
     });
+
+    this.mapService.loadPath$.pipe(
+      takeUntil(this.destroy$),
+      skip(1)
+    )
+    .subscribe({
+    next: (pathData: { plate: string, points: Point[] }) => {
+      const startPoint: number = pathData.points[0].lat;
+      const endPoint: number = pathData.points[0].long;
+
+      this.initMap(new Point(startPoint, endPoint));
+
+      // Remove previous routing control if it exists
+      if (this.routingControl) {
+          this.map.removeControl(this.routingControl);
+          this.routingControl = null;
+      }
+
+
+      const waypointsCompleti = pathData.points.map((point) =>
+          L.latLng(point.lat, point.long)
+      );
+
+      const waypointsVisibili = [
+        waypointsCompleti[0],
+        waypointsCompleti[waypointsCompleti.length - 1]
+      ];
+
+      waypointsVisibili.forEach((waypoint, index) => {
+        const newMarker = L.marker(waypoint);
+        let popupContent = "";
+        if (index === 0) {
+            popupContent = "Inizio";
+        } else if (index === waypointsVisibili.length - 1) {
+            popupContent = "Fine";
+        }
+
+        try {
+            newMarker.addTo(this.map)
+                .bindPopup(
+                  this.mapService.getCustomPopup(popupContent),
+                  {
+                    autoClose: false
+                  }
+                )
+                .openPopup();
+        } catch (error) {
+            console.error("Error creating marker:", error);
+        }
+      });
+
+      this.routingControl = L.Routing.control({
+          waypoints: waypointsCompleti,
+          routeWhileDragging: false,
+          addWaypoints: false,
+      }).addTo(this.map);
+
+      this.routingControl.setWaypoints(waypointsVisibili);
+
+      //Forza il ricalcolo del percorso. Utile in alcuni casi
+      this.routingControl.route();
+
+    },
+    error: error => console.error("Errore nel caricamento del percorso: ", error)
+  });
+
+
   }
 
-  private initMap() {
+  /**
+   * Inizializza la mappa su un punto
+   * @param point punto da cui inizializzare la mappa
+   */
+  private initMap(point: Point) {
     this.initialized = true;
     this.cd.detectChanges();
-    this.mapService.removeMap(this.map);
-    this.map = this.mapService.initMap(this.map, this.lat, this.long);
+    if(this.map) this.mapService.removeMap(this.map);
+    this.map = this.mapService.initMapByPoint(this.map, point);
   }
+
 }
